Organize components by page
