+++ src/components/common/Hero.css
