+++ src/pages/index.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import Layout from '@theme/Layout';
-import HomepageFeatures from '../components/HomepageFeatures';
-import FreeSoftware from '../components/FreeSoftware';
-import BuiltByVolunteers from '../components/BuiltByVolunteers';
-import MoreClients from '../components/MoreClients';
-import CallToAction from '../components/CallToAction';
-import Hero from '../components/Hero';
+
+import HomepageFeatures from '../components/home/HomepageFeatures';
+import FreeSoftware from '../components/home/FreeSoftware';
+import BuiltByVolunteers from '../components/home/BuiltByVolunteers';
+import MoreClients from '../components/home/MoreClients';
+import CallToAction from '../components/home/CallToAction';
+import Hero from '../components/common/Hero';
 
 export default function Home() {
   return (
